% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan.R
\name{MiXcan}
\alias{MiXcan}
\title{Estimate the cell-type level prediction weights of a gene.}
\usage{
MiXcan(y, x, cov = NULL, pi, xNameMatrix = NULL, yName = NULL, foldid = NULL)
}
\arguments{
\item{y:}{The pre-cleaned expression level data for a single gene in N samples.}

\item{x:}{A N by P matrix for all the genetic predictors used to predict the genetically regulated expression  of the gene.}

\item{cov:}{A N by Q matrix for the covariates adjusted in the model (e.g. age, population stratification).}

\item{pi:}{An estimation of cell-type fraction for the cell type of interest (e.g.
epithelial). It can be obtained using existing methods
in the literature or from the output of pi_estimation function.}

\item{xNameMatrix:}{Default is NULL. A matrix to save theX matrix information,
such as variable ID, position, rsid, ref_allele, eff_allele.}

\item{yName:}{Default is NULL. A row vector to save the expression information, such as gene ID, gene name.}

\item{foldid:}{Default is NULL. 10-fold cross-validation (CV) is used in our pipeline. A random split
is considered if foldid is NULL. Otherwise foldid is used to split the data for CV.}
}
\value{
list with 9 elements. It contains

\item{type:}{Whether the prediction model is "CellTypeSpecific" or "NonSpecific".}
\item{beta.SNP.cell1:}{The prediction weights of the genetic predictors in cell type 1 (the cell type of interest).}
\item{beta.SNP.cell2:}{The prediction weights of the genetic predictors in cell type 2 (other cell types).}
\item{beta.all.models:}{All regression coefficients are saved in beta.all.models, including intercepts,
coefficients of genetic and non-genetic predictors in cell-type specific and non-specific models.}
\item{glmnet.cell:}{The cell-type-level prediction model selected using elastic-net.
This model may not be the final model as elastic-net selected parameters in the two
cell types may not be robustly different.  }
\item{glmnet.tissue:}{The tissue-level prediction model, which does not consider
cell type composition (same as PrediXcan).}
}
\description{
The core function of MiXcan package for estimating the
cell-type level prediction weights of a gene.
}
