% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan_association.R
\name{MiXcan_association}
\alias{MiXcan_association}
\title{Association analysis with MiXcan predicted gene expression levels}
\usage{
MiXcan_association(MiXcan_predicted_expr, covariates, outcome, family)
}
\arguments{
\item{MiXcan_predicted_expr:}{A N by 2 matrix indicating the predicted gene expression levels in two
cell types. If a non-specific model is used for prediction, the predicted values should be the same in two cell types.}

\item{covariates:}{A N by P matrix for the covariates in association analysis.}

\item{outcome:}{The dependent variable of the association analysis. It can be continuous, categorical
and count variables, and modeled through generalized linear models (GLM).}

\item{family:}{Same as the family object used in GLM.}
}
\value{
A tibble showing the estimate, standard error, p value for cell 1 and 2 and the combined p value using the Cauchy combination.
}
\description{
Association analysis with MiXcan predicted gene expression levels
}
