% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan_association.R
\name{MiXcan_association}
\alias{MiXcan_association}
\title{Association analysis with MiXcan predicted genetically regulated expression}
\usage{
MiXcan_association(new_y, new_cov, new_outcome, family = gaussian)
}
\arguments{
\item{new_y:}{A N by 2 matrix indicating the predicted gene expression levels in two
cell types. If a non-specific model is used for prediction, the predicted values should be the same in two cell types.}

\item{new_cov:}{A N by P matrix for the covariates in association analysis.}

\item{new_outcome:}{The dependent variable of the association analysis. It can be continuous, categorical
and count variables, and modeled through generalized linear models (GLM).}

\item{family:}{Same as the family object used in GLM.}
}
\value{
A tibble showing the estimate, standard error, p value for cell 1 and 2
and the combined p-value using the Cauchy combination.
}
\description{
The input MiXcan predicted genetically regulated expression levels are at cell-type
level. Association analysis returns p-value at tissue level for inference purpose, and
at cell-type level for post-inference (ad hoc) explaination of the likely source of
association.
}
